refactor: Navbar 구현 및 관심사 분리
