+++ src/layouts/MobileView.tsx
@@ -1,21 +1,9 @@
 import styled from '@emotion/styled';
-import {
-  Navigation,
-  SVGNavActiveFeed,
-  SVGNavFeed,
-  SVGNavActiveProfile,
-  SVGNavProfile,
-  SVGWrite24,
-} from 'concept-be-design-system';
-import { Outlet, useNavigate } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 
-import useRouteMatched from '../hooks/useRouteMatch';
+import Navbar from './Navbar';
 
 const MobileView = () => {
-  const { hasMatched } = useRouteMatched();
-  const navigate = useNavigate();
-  const isMatchedNavigation = hasMatched('/profile/:id', '/login', '/write', '/agreement', '/auth', '/feed/:id');
-
   // TODO: Header 도메인 얽힘 문제 해결 확인 시 사용 예정
   // const isMatchedHeader = hasMatched('/feed', '/feed/:id', '/profile', '/profile/:id', '/profile/:id/more');
 
@@ -24,19 +12,7 @@ const MobileView = () => {
       <Wrapper>
         <Outlet />
       </Wrapper>
-      {!isMatchedNavigation && (
-        <Navigation>
-          <Navigation.Item onClick={() => navigate('/feed')}>
-            {location.pathname.startsWith('/feed') ? <SVGNavActiveFeed /> : <SVGNavFeed />}
-          </Navigation.Item>
-          <Navigation.Item position="center">
-            <SVGWrite24 onClick={() => navigate('/write')} />
-          </Navigation.Item>
-          <Navigation.Item onClick={() => navigate('/profile')}>
-            {location.pathname.startsWith('/profile') ? <SVGNavActiveProfile /> : <SVGNavProfile />}
-          </Navigation.Item>
-        </Navigation>
-      )}
+      <Navbar />
     </Container>
   );
 };
